fix: update assertions expecting  in test to use int instead to ensure that data returned from database match with the expected test value

+++ app/src/test/java/com/littlelemon/application/address/data/DefaultAddressRepositoryTest.kt
@@ -407,7 +407,7 @@ class DefaultAddressRepositoryTest {
         val addressCount = repository.getAddressCount()
 
         // Then, five is returned
-        assertEquals(expectedAddressCount.toLong(), addressCount)
+        assertEquals(expectedAddressCount, addressCount)
     }
 
     @Test
